charaCli/requirements.txt
beautifulsoup4==4.7.1
certifi==2020.6.20
chardet==3.0.4
idna==2.8
pychara==0.9.3
python-dotenv==0.14.0
requests==2.21.0
soupsieve==2.0.1
texttable==1.6.2
tqdm==4.48.2
urllib3==1.24.3
